requirements.txt
altair==4.2.2
altgraph==0.17.3
anyio==3.6.2
appdirs==1.4.4
astroid==2.12.13
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.16.3
firebase==4.0.1
firebase-admin==6.2.0
Flask==2.2.2
fonttools==4.38.0
forex-python==1.8
frozendict==2.3.8
future==0.18.2
gcloud==0.17.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.1
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
html5lib==1.1
httpcore==0.16.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==6.6.0
influxdb==5.3.1
ipykernel==6.22.0
ipython==8.13.1
ipython-genutils==0.2.0
isort==5.11.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jws==0.1.3
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
Levenshtein==0.20.8
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
msgpack==1.0.4
multitasking==0.0.11
nest-asyncio==1.5.6
numpy==1.23.5
oauth2client==3.0.0
openpyxl==3.0.10
orjson==3.8.3
packaging==22.0
pandas==1.5.2
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
pydantic==1.10.2
pydeck==0.8.1b0
Pygments==2.15.1
pyinstaller==5.5
pyinstaller-hooks-contrib==2022.10
PyJWT==2.6.0
pylint==2.15.9
Pympler==1.0.1
pymssql==2.2.7
pyodbc==4.0.35
pyparsing==3.0.9
pyperclip==1.8.2
Pyrebase==3.0.27
pyrsistent==0.19.3
python-arango==7.5.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jwt==2.0.1
python-Levenshtein==0.20.8
python-multipart==0.0.5
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.3.1
rapidfuzz==2.13.6
requests==2.31.0
requests-toolbelt==0.7.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
Send2Trash==1.8.2
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
stack-data==0.6.2
starlette==0.22.0
streamlit==1.24.0
tenacity==8.2.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
ujson==5.6.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.20.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
yapf==0.32.0
yfinance==0.2.18
zipp==3.15.0
